Move blueprints to blueprints folder, move database.py to app folder
