+++ queertk/blueprints/admin/templates/customview.html
