+++ queertk/blueprints/authentication/models.py
@@ -1,4 +1,4 @@
-from database import db
+from queertk.database import db
 from flask_login import UserMixin
 from werkzeug.security import generate_password_hash, check_password_hash
